add proper handle for non authorized users

+++ common/domain/src/commonMain/kotlin/work/racka/reluct/common/domain/usecases/tasks/impl/ManageTaskLabelsImpl.kt
@@ -14,21 +14,32 @@ import work.racka.reluct.common.domain.mappers.tasks.asTaskLabel
 import work.racka.reluct.common.domain.mappers.tasks.asTaskLabelDbObject
 import work.racka.reluct.common.domain.usecases.tasks.ManageTaskLabels
 import work.racka.reluct.common.model.domain.tasks.TaskLabel
+import work.racka.reluct.common.model.enums.responses.NetworkPushResponse
+import work.racka.reluct.common.network.sync.tasks.TasksUpload
 import work.racka.reluct.common.services.haptics.HapticFeedback
 
 internal class ManageTaskLabelsImpl(
     private val dao: TasksDao,
+    private val tasksUpload: TasksUpload,
     private val haptics: HapticFeedback,
     private val dispatcher: CoroutineDispatcher = Dispatchers.IO
 ) : ManageTaskLabels {
 
     override suspend fun addLabel(label: TaskLabel) = withContext(dispatcher) {
-        dao.addTaskLabel(label = label.asTaskLabelDbObject())
+        val mapped = label.asTaskLabelDbObject()
+        val resp = tasksUpload.uploadLabel(mapped)
+        if (resp !is NetworkPushResponse.Success) {
+            dao.addTaskLabel(label = label.asTaskLabelDbObject())
+        }
         haptics.click()
     }
 
     override suspend fun addLabels(labels: ImmutableList<TaskLabel>) = withContext(dispatcher) {
-        dao.addAllTaskLabels(labels = labels.map { it.asTaskLabelDbObject() })
+        val mapped = labels.map { it.asTaskLabelDbObject() }
+        val resp = tasksUpload.uploadLabels(mapped)
+        if (resp !is NetworkPushResponse.Success) {
+            dao.addAllTaskLabels(labels = labels.map { it.asTaskLabelDbObject() })
+        }
         haptics.click()
     }
 
@@ -43,11 +54,15 @@ internal class ManageTaskLabelsImpl(
         .flowOn(dispatcher)
 
     override suspend fun deleteLabel(id: String) = withContext(dispatcher) {
-        dao.deleteTaskLabel(id)
+        val resp = tasksUpload.deleteLabel(id)
+        if (resp !is NetworkPushResponse.Success) {
+            dao.deleteTaskLabel(id)
+        }
         haptics.heavyClick()
     }
 
     override suspend fun deleteAllLabels() = withContext(dispatcher) {
+        // TODO: Implement on Firebase too
         dao.deleteAllTaskLabels()
         haptics.heavyClick()
     }
